Moved a few more methods/constants out of calcVarPriors into a new module, dataproc, that deals with formatting and setting defaults for the input variants.
